Extract mojarra namespace to its own typed module #5506
