+++ impl/src/main/ts/faces.ts
@@ -3054,166 +3054,3 @@ if ( !( (window.faces && window.faces.specversion && window.faces.specversion >=
 
 
 } //end if version detection block
-/*
- * Copyright (c) 1997, 2020 Oracle and/or its affiliates. All rights reserved.
- * Copyright 2004 The Apache Software Foundation
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-/**
- @project Faces Ajax Library
- @version 2.0
- @description This is the standard implementation of the Faces Ajax Library.
- */
-
-/**
- * @name mojarra
- * @namespace
- */
-
-/*
- * Create our top level namespaces - mojarra
- */
-window.mojarra = window.mojarra || {};
-
-
-/**
- * This function deletes any hidden parameters added
- * to the form by checking for a variable called 'adp'
- * defined on the form.  If present, this variable will
- * contain all the params added by 'apf'.
- *
- * @param f - the target form
- */
-mojarra.dpf = function dpf(f) {
-    const adp = f.adp;
-    if (adp !== null) {
-        for ( const param of adp ) {
-            param.remove();
-        }
-    }
-};
-
-/**
- * This function adds any parameters specified by the
- * parameter 'pvp' to the form represented by param 'f'.
- * Any parameters added will be stored in a variable
- * called 'adp' and stored on the form.
- *
- * @param f - the target form
- * @param pvp - associative array of parameter
- *  key/value pairs to be added to the form as hidden input
- *  fields.
- */
-mojarra.apf = function apf(f, pvp) {
-    const adp = [];
-    f.adp = adp;
-    let i = 0;
-    for (const k of Object.keys(pvp) ) {
-        const p = document.createElement("input");
-        p.type = "hidden";
-        p.name = k;
-        p.value = pvp[k];
-        f.appendChild(p);
-        adp[i++] = p;
-    }
-};
-
-/**
- * This is called by command link and command button.  It provides
- * the form it is nested in, the parameters that need to be
- * added and finally, the target of the action.  This function
- * will delete any parameters added <em>after</em> the form
- * has been submitted to handle DOM caching issues.
- *
- * @param f - the target form
- * @param pvp - associative array of parameter
- *  key/value pairs to be added to the form as hidden input
- *  fields.
- * @param t - the target of the form submission
- */
-mojarra.cljs = function cljs(f, pvp, t) {
-    mojarra.apf(f, pvp);
-    const ft = f.target;
-    if (t) {
-        f.target = t;
-    }
-
-    const input = document.createElement('input');
-    input.type = 'submit';
-    f.appendChild(input);
-    input.click();
-    input.remove();
-
-    f.target = ft;
-    mojarra.dpf(f);
-};
-
-/**
- * This is called by functions that need access to their calling
- * context, in the form of <code>this</code> and <code>event</code>
- * objects.
- *
- *  @param f the function to execute
- *  @param t this of the calling function
- *  @param e event of the calling function
- *  @return object that f returns
- */
-mojarra.facescbk = function facescbk(f, t, e) {
-    return f.call(t,e);
-};
-
-/**
- * This is called by the AjaxBehaviorRenderer script to
- * trigger a faces.ajax.request() call.
- *
- *  @param s the source element or id
- *  @param e event of the calling function
- *  @param n name of the behavior event that has fired
- *  @param ex execute list
- *  @param re render list
- *  @param op options object
- */
-mojarra.ab = function ab(s, e, n, ex, re, op) {
-    if (!op) op = {};
-    if (n)   op["jakarta.faces.behavior.event"] = n;
-    if (ex)  op["execute"] = ex;
-    if (re)  op["render"] = re;
-    faces.ajax.request(s, e, op);
-};
-
-/**
- * This is called by command script when autorun=true.
- *
- * @param l window onload callback function
- */
-mojarra.l = function l(l) {
-    if (document.readyState === "complete") {
-        setTimeout(l);
-    }
-    else {
-        window.addEventListener("load", l);
-    }
-};
-
-/**
- * Add event listener.
- *
- * @param id element id
- * @param ev event name
- * @param fn function
- */
-mojarra.ael = function ael(id, ev, fn) {
-    document.getElementById(id).addEventListener(ev, fn);
-};
